requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.1
pandas==2.2.2
mysql-connector-python==8.1.0
SQLAlchemy==2.0.19
